Factor out common socket class logic
